Do not use a file with a predictable name in /tmp
Using a file with a predicatable name in /tmp is likely a source of trouble on a shared system, since /tmp/ may be shared among all users. /run/user/ is made for that, cleaned on reboot, and properly protected from interference from others. It requires systemd and a newer system.